Safari-609.1.11.1/LayoutTests/platform/mac/imported/w3c/web-platform-tests/2dcontext/line-styles/2d.line.cap.round-expected.txt

2d.line.cap.round
lineCap 'round' is rendered correctly
Actual output:
Expected output:


FAIL lineCap 'round' is rendered correctly assert_equals: Red channel of the pixel at (67, 6) expected 0 but got 1
